Added data access operator for chunks
I have not yet given any thought to handling the different coordinate
systems. If I am to organize chunks in a, well, for want of a better
word, `ChunkManager`, the _external_ coordinates need to be converted
into the internal chunk coordinates. That way, no chunk knows _anything_
about its relation to the world.

Not sure whether this is a good stance to take at this time.